// PensionManagementUI/src/app/features/pages/all-details/models/pensiondetail.model.ts
export interface PensionDetails{
    pensionerId: number;
    fullName: string;
    dateOfBirth: string;
    gender: string;
    aadharNumber: string;
    phoneNumber: string;
    address: string;
    age: number;
    id: string;
    pensionPlanId: string
    pensionPlanDetails:{

        pensionPlanName:''

    }
}